open_main_menu_page() in main
